requirements.txt
Flask==3.0.0
flask-restx==1.3.0
seedrcc
python-dotenv==1.0.0
werkzeug==3.0.1
